REFACTOR: MyRendHistory -> MyRentInfo
